move internal shadow widget to internal widget package
